LIBITD-2422. Updated config to use "sibling" reciprocal-borrowing app
Updated the configuration to enable the "reciprocal-borrowing"
application to be checked out into a sibling directory of
this repository, instead of within this repository.

Pinned the IdP to 4.3.1, to prevent any unexpected changes.

When running the Reciprocal Borrowing application, modified the
Passenger/Apache configuration in "00-virtualhosts.conf" to use
the "development_docker" Rails environment.

Updated the README.md with information about using the development
environment.

https://umd-dit.atlassian.net/browse/LIBITD-2422