LIBITD-2589. Tweak files for Rails 8.1
Add "yarn" and "nodemon" to the "sp/Dockerfile" because they are used
by the Rails asset pipeline to provide dynamic updates for the CSS.

Updated the "README.md" to include running "yarn install" and
"yarn watch:css" to support local development.

Updated "00-virtualhosts.conf" to use the Rails "development"
environment, because the previous "development_docker" environment
was removed (it was not compatible with new asset pipeline).

Tweaked the "startup.sh" script to add a "/run/shibboleth" directory,
as that appears to be necessary for the Shibboleth SP to start properly.

https://umd-dit.atlassian.net/browse/LIBITD-2589